FIXED: minior bugs
When ever the id was retrieved from the database it was returned as an array and it was being treated as an integer. Created a method to convert a single value array to just an integer. Also, the Tag object was missing an attribute assignment of date in the initialization function so the date the tag was created was not being stored in the database. With this a successful retreival of the history of the given repository was done.