DDED: file parser now parses a protect's files
It will now go through an parse the repo's files. It also now stores important counts as well as the actual documentation and code (seperately). Finally, it stores those values for each version of the file in the repo. (over the life of the project). Started fixin up the github scraper to pull the types of files in the project, python, ruby, or java.