ADDED: file parser for java
Since I am more confident with java and the comment documentation standards in java I've switched the methods to start using the regex for java. Fixed an issue with collecting files that were originally windows projects with the carriage return. Also, added a list of commenting styles and examples of each of them.