ADDED: package level querying
Packages can now be views, not all of them yet only packages with direct children that are java files. Also, fixed an issue with the the total code and comments for large amounts of data. Since it was previously show as a sum if the values could not fit on screen, now the average is shown which is far more accurate.