FIXED: some issues with quotations and reseting
Previously no check was made to ensure a comment start flag that was part of a quote (ex. " a starter /* is used .."). A check is now done on the the code. Also, the array of lines was not reseted each time for the patch thus if there was more than 1 patch for a file it would overlap.