ADDED: low and high threshold and store more info
The modification threshold is now a certain percentage for lines under a certain length and another precentage for lines equal or longer than that length (all of these are can be defined). Also, added the new schema and accessor/setter methods for collecting the tags and user information for the projects.