FIXED: quote removal and more minor bugs
Quote removal only needs to be single line quote removal. Since it was not working as is, it was similified to work correctly. Also method finder is having issues the merged files. Since the file contains both added and deleted code the method finding can very easily break (- method then + method will make so the method may not end properly). The current fix for that was ignoring lines with a minus (for just the method_finder). Methods with a throws were missed so regex was fixed. Finally, anon. class delcaration were captured as methods and were fixed accordingly.