MODIFIED: comment removal and comment detection
Added back in the comment removal to method finder. Also, added the comment start detection. A comment for a method, that is a comment preceeding the method is detected. By using the start of the comment to the 'actual start' of the method signature the length of the comment can be calculated and extracted accordingly. Currently deleted lines are deleted since an edit to the method signature (delete and addition) will cause the merge file to not have a correctly structured file.