MODIFIED: information collection
Improved some of the basic database interface methods. Added a method for fetching the results. Also, added a method to return an associated array of values. The method takes the column names (as defined in mysql or in the query as the key and assigns the value to it in a hash. Finally, add an output file for significant method changes.