FIXED: merger bug
The full file (file + patch) was not correctly aligned. Since deleted statements are added into the code (preceeded by a '-') the indexes of the patch headers are incorrect. This was accounted for by keeping track of the number of deleted lines to allow for the patch to be offset. However, the deletions were reset after a new patch header was found. This would lead to cases where the additions were completely off. The case would be where two patchs contained deletions. This would then throw off the patch header line count as well as the correction mechanism. A simple fix was applied, make so that a new patch header does not reset the deletion count.