FIXED: incorrect patches
Two types of patches were corrected. In the case were the entire file is deleted the patch would merge the deletions into a un-marked up version of the file (duplicating lines). This was fixed so no only the deleted marked up file is presented.

The second case is when several patches are provided that contain deletions. The line count did not take into consideration previous patch deletions and therefore the line offset correct was incorrect and lead to misaligned patch files.