FIXED: delete method identification
Several issues surrounded detecting deleted methods. Previously the delete method signatures were ignored to prevent invalid depth detection. However this was to aggresive since it prevented completely deleted methods from being detected. Currently the conditions are a bit more relaxed. However, in some cases methods will be now picked up as having erronious ends. This would be in the case were the method signature change type is different from that of the end curly brace. This typically happens when the method signature is unchanged but the end curly brace is deleted.