Changed the method signature to method id
The plot originally had the full method signature, in order to make it a little easier to line up it was limited to a maximum of 10 characters. However that really didn't tell much of anything. So now the method signature only shows the method's first method_id (from the database).