Added: analysis for using svm.
SVM tester will use the extracted data (from get_data.rb) to train an svm model and test against another set of data (that occurs in the future). This allows for validating the svm model created. get_data uses a pg-interface (ported from mysql to allow for more advanced functions). Finally, quarter_tester allows for stripped testing of the data to be done.